change: change integration entity structure
